[Build] Improve Docker build instructions
The work done so far as part of #629 means that we no longer require a
Python (virtual) environment to install all components of OpenAssetIO.
We can simply `cmake --install` and both the compiled C++ artifacts and
the pure Python source artifacts will be bundled under the
`CMAKE_INSTALL_PREFIX` directory.

So simplify the Docker build instructions to do this instead of using
the `openassetio-python-deploy` target.

This has the added advantage of not leaving an unusable (outside of the
Docker container) Python environment on the host system, mixed in with
the installed artifacts, meaning the `dist` directory created using the
Docker one-liner is theoretically relocatable to any compatible
platform.

Hence delete the defunct `openassetio-python-deploy` target.

Signed-off-by: David Feltell <david.feltell@foundry.com>