[Build] Allow test execution without venv
Part of #629. Outside of the core dev team it is more typical to add
packages to an external Python environment. We did not previously
support running the unit tests in this situation, without significant
manual steps.

Add a new `OPENASSETIO_ENABLE_PYTHON_TEST_VENV` CMake option, `ON` by
default. If disabled, then tests will (attempt to) run in the external
Python environment. If enabled, then a Python environment will be
created as a test fixture, as before.

Add a new preset `test-custom-python-env` that is identical to the
`test` preset, but disables Python the venv creation fixture.

As the python sources are no longer installed into the python
environment, `PYTHONPATH` must be extended to cover the install tree.
Users may find it burdensome to manually have to extend their
`PYTHONPATH` when testing, so automatically extend the `PYTHONPATH`
appropriately for the test.

Move the test venv out of `build/dist` and into `build/test-venv`,
which has the benefit of making the `dist` folder a clean bundle
directory.

Removed unnecesary `--no-build-isolation` flag from pip install when
executed in cmake, as this is redundant due to no longer performing
python packaging via cmake. Similarly, removed `wheel` from the test
virtual environment, as that was a dependency for
`--no-build-isolation`.

Also removed step-by-step section in building documentation, as it was
incorrectly referencing pytest rather than ctest, and is now even
further redundant. #624 is slated to rewrite this document.

Signed-off-by: David Feltell <david.feltell@foundry.com>