[Build] Add CIBuildWheel Actions Pipeline
Part of #653

In order to have an automated release pipeline to provide prebuilt
wheels to PyPI, establish an automated pipeline that builds python
wheels across multiple platforms/python versions, using cibuildwheel.

As of this commit,  targets python3.9 and python3.10, on Windows
and Ubuntu, x86_64 only. (Matrix of 4).

Pipeline will be invoked on every push to main branch.

Implementation note : This commit includes a change to migrate the conan
cmake toolchain file path from user home to the repository. This was
done to provide consistency with non-CI build steps, as well as to allow
cibuildwheel (or more correctly, CMake invoked under the docker
container cibuildwheel uses on Linux platforms) to simply accept
the file argument as a relative path.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>