[Build] Use consistent conan location
Part of #653

Changes the conan install directory in all of our tooling
to consistently be `{{github.workspace}}/.conan`

In order to have consistent expansion of paths across multiple
platforms, also replaces inconsistent usages of $WORKSPACE with
`${{github.workspace}}`

Implementation note : This commit includes a change to migrate the conan
cmake toolchain file path from user home to the repository. This was
done to provide consistency with non-CI build steps, as well as to allow
cibuildwheel (or more correctly, CMake invoked under the docker
container cibuildwheel uses on Linux platforms) to simply accept
the file argument as a relative path.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>