[Build] Remove vendored third party package
Make it so OpenAssetIO dosent rely on any vendored thirdparty packages.
Previously, we had a version of toml++ vendored, remove that, and
instead have it fetched from conan/whatever dependency resolution system
the user is using.

Consequently, we need to create a new docker image beyond the CY2022
reference image, as that dosent contain toml++.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>