[Core] Add python conversion functions
[#798] There is a need as a hybrid C++/Python host application author
to convert types between an embedded python interpreter and a c++
context, as some hosts will want to do asset resolution in both
languages, sharing the same configuration.

This adds two methods to python bridge, `castToPyObject`
and `castFromPyObject` that facilitates this.

Prior, the way to do this was to do a pybind cast, which
is what we do here behind the scenes, but by providing
methods to raw CPython types, we eliminate the need
to link against pybind for the consumer.

Also adds an additional CTest project,
`openassetio.internal.python-bridge-no-openassetio-module-test` that
runs tests tagged with `[no_openassetio_module]`, on order to guarantee
test runs without the openassetio having already been loaded/linked.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>