[Test] Pybind `get_local_internals()` exception
Pybind employs the common technique of initializing a static variable
with `new` and never cleaning it up, ie, letting it leak.

This code-path is only triggered when a py::cast is not performed due
to there not being a type registered to cast into, hence why the
sanitizers do not report a memory leak normally, but we do get one in
the case we are testing where `openassetio` isn't imported to the python
interpreter.

See `pybind11::detail::type_caster_base::src_and_type`.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>