[Core] Add Pager template object
Add the pager and pager interface template object, to support interfaces
that may return large datasets.

Pager and PagerInterface follow a similar pattern to Manager and
ManagerInterface, with PagerInterface intended as a base interface
for a manager to implement, with Pager being the object that the host
constructs.

Explicitly instantiated to `EntityReference`, as the only currently
slated paginated method is `getRelatedReferences`, which is of the
EntityReference type. The intent being that when that method is
implemented, an alias to the template object will come with it.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>