[Core] Refine Context.access modes
Adds the `kCreateRelated` context access mode.
This is an access mode necessary to signal intent to create related
entities, rather than mutations/versions of an entity. This comes up
particularly in the "folder in a folder" case, where it's very likely
that the regular method of determining version/new, (checking that the
traitsets match,) will not be enough.

Also remove the idea of multiple writes, due to ther being no coherent
use for these within the core API methods themselves, and so this
behavior could be more accurately and coherently expressed as part of
the UI layer itself.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>