[Core] Optional exception data.
In order to handle cases where the manager may emit
an unexpected BatchElementError in a convenience,
make all data in the Batch exceptions optional.

This is one extreme of the permissive/strict binary, favouring
permissiveness, ensuring hosts get the same error type the
manager intended, even if it doesen't come with the data.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>