[CI] Publish release artifacts.
Publish build artifacts as assets on release page #1201

Extend the build matrix of our mainline build pipeline script to build
appropriate variants for release artifacts. Then upload them on a
release.

This is good for downstream users, but will also allow us to pin to
releases in our own CI easier, in the case we need C++ versions.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>